refactor: move PeerSession to network crate
also move Peer to kernel as it's a cross-cutting concern
