+++ crates/amaru-network/Cargo.toml
@@ -17,6 +17,8 @@ anyhow.workspace = true
 hex.workspace = true
 pallas-crypto.workspace = true
 pallas-math.workspace = true
+pallas-network.workspace = true
+pallas-traverse.workspace = true
 rayon.workspace = true
 serde.workspace = true
 thiserror.workspace = true
